+++ Student-Course-ManagmentSystem/Repository/AssignStudentRepositories/AssignStudent.cs
@@ -3,10 +3,7 @@ using Student_Course_ManagmentSystem.ControllerModels;
 using Student_Course_ManagmentSystem.DataModels;
 using System;
 using System.Collections.Generic;
-using System.Text.Json.Serialization;
-using System.Text.Json;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
 using System.Linq;
 
 namespace Student_Course_ManagmentSystem.Repository.RegisterRepositories
@@ -25,11 +22,11 @@ namespace Student_Course_ManagmentSystem.Repository.RegisterRepositories
             {
                 StudentName = g.Student.Name,
                 CourseName = g.Course.Name,
-                GradeValue = (int)g.Score
+                GradeValue = (decimal)g.Score
             }).ToList();
             return gradeDTOs;
         }
-        public async Task<string> RegisterStudent(string studentId, string courseId, int grade)
+        public async Task<string> RegisterStudent(string studentId, string courseId, decimal grade)
         {
             var registration = new Grade { StudentId = Guid.Parse(studentId), CourseId = Guid.Parse(courseId), Score = grade };
             await db.Grades.AddAsync(registration);
